feat(transactions): TransactionRowView gains overdue / due-today / Pay parameters
Adds four optional parameters to the canonical row so it can render the
special UX previously inlined in `UpcomingTransactionRow`:

- `isOverdue: Bool = false` — red `exclamationmark.triangle.fill` icon
  + red payee text.
- `isDueToday: Bool = false` — orange + bold date in the meta row.
- `onPay: (() -> Void)?` — inline trailing "Pay" button when non-nil.
- `pendingPayId: Transaction.ID?` — when equal to this row's id, the
  Pay button is replaced by a `ProgressView` with a payee-parameterised
  `.accessibilityLabel` and the row is `.disabled(true)`.

Recurrence description (`transaction.recurPeriod` + `recurEvery`) is
also added to the meta row when present.

All existing call sites unchanged — every new parameter defaults to a
behaviour-preserving value.

The preview helpers are extracted to `TransactionRowView+Preview.swift`
to keep the main file under SwiftLint's 400-line file_length warning.
